alpha-backend/requirements.txt
setuptools>=65.0.0
fastapi==0.115.4
uvicorn==0.32.0
python-multipart==0.0.12
qdrant-client==1.12.1
PyMuPDF==1.24.11
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0
openai==1.54.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
numpy==1.26.4
sentence-transformers==2.7.0
scikit-learn==1.4.2
torch>=2.2.0

